% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{munster}
\alias{munster}
\title{Cycling network of Münster, Germany}
\format{
A \code{cycling_network} object with components:
\describe{
\item{city}{Character. Name of the city.}
\item{lines}{An \code{sf} object with LINESTRING geometries.}
\item{download_date}{Date of download.}
}
}
\source{
OpenStreetMap contributors \url{https://www.openstreetmap.org}
}
\usage{
munster
}
\description{
A \code{cycling_network} object downloaded from OpenStreetMap
containing the cycling infrastructure of Münster, Germany.
}
\keyword{datasets}
