% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycling_network.R
\name{plot.cycling_network}
\alias{plot.cycling_network}
\title{Plot a cycling_network instance}
\usage{
\method{plot}{cycling_network}(x, ...)
}
\arguments{
\item{x}{A \code{cycling_network} instance.}

\item{...}{Further arguments (currently ignored).}
}
\value{
A \code{ggplot} object, invisibly.
}
\description{
Produces a \code{ggplot2} map of the raw cycling network without classification.
}
\examples{
\dontrun{
net <- get_cycling_network("Muenster, Germany")
plot(net)
}

}
