% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycling_classification.R
\name{print.cycling_classification}
\alias{print.cycling_classification}
\title{Print a cycling_classification instance}
\usage{
\method{print}{cycling_classification}(x, ...)
}
\arguments{
\item{x}{A \code{cycling_classification} instance}

\item{...}{Further arguments (currently ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Displays the city, download date, number of classified segments and the infrastructure type summary table.
}
\examples{
\dontrun{
cl <- classify_bike_infrastructure(get_cycling_network("Muenster, Germany"))
print(cl)
}

}
