% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycling_network.R
\name{print.cycling_network}
\alias{print.cycling_network}
\title{Print a cycling_network instance.}
\usage{
\method{print}{cycling_network}(x, ...)
}
\arguments{
\item{x}{A \code{cycling_network} instance.}

\item{...}{Further arguments (currently ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Displays a short summary of the cycling network instance to the console.
}
\examples{
\dontrun{
net <- get_cycling_network("Muenster, Germany")
print(net)
}

}
